requirements.txt
requests
beautifulsoup4
lxml
fpdf2
streamlit
